logger changes

+++ Java/src/Iron_Dome.java
@@ -1,6 +1,7 @@
 import java.io.IOException;
 import java.util.Calendar;
 import java.util.logging.FileHandler;
+import java.util.logging.Level;
 
 public class Iron_Dome extends Thread {
 	private String id;
@@ -12,36 +13,28 @@ public class Iron_Dome extends Thread {
 		this.id = "D" + (int) (Math.random() * 1000);
 	}
 
-	public Iron_Dome(String id) throws InterruptedException {
+	public Iron_Dome(String id) throws Exception {
 		this.id = id;
-		FileHandler theHandler;
-		try {
-			theHandler = new FileHandler("IrodDome" + id + ".txt");
-			theHandler.setFilter(new ObjectFilter(this));
-			theHandler.setFormatter(new LogFormatter());
-			War.theLogger.addHandler(theHandler);
-		} catch (SecurityException e) {
-			e.printStackTrace();
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
+		War.theLogger.addHandler((new Handler(this.getClass().getName(), id, this)));
 	}
 
-	public void checkIfPossibleToIntercept(String destructAfterLaunch, String id)
-			throws InterruptedException {
+	public void checkIfPossibleToIntercept(String destructAfterLaunch, String id) throws InterruptedException {
 		int destruct_After_Launch = Integer.parseInt(destructAfterLaunch);
 		for (Enemy_Missile enemy_missile : War.enemyMissile) {
 			if (enemy_missile.getID().equalsIgnoreCase(id)) {
 				synchronized (enemy_missile) {
-					if (!(destruct_After_Launch >= enemy_missile.getFlyTime() + enemy_missile.getLaunchTime()))
-					{
-						System.out.println(Calendar.getInstance().getTime()
-								+ "\t Iron dome #" + this.id
-								+ " is hitting enemy missile #" + id + " in "
-								+ destruct_After_Launch + " sec ");
+					if (!(destruct_After_Launch >= enemy_missile.getFlyTime() + enemy_missile.getLaunchTime())) {
+						if (enemy_missile.isAlive)
+							War.theLogger.log(Level.INFO, "Iron dome #" + this.id + " attempting to intercept missile #" + enemy_missile.getID() + " in " + destruct_After_Launch + " sec ", this);
+
+						System.out.println(Calendar.getInstance().getTime() + "\t Iron dome #" + this.id + " is hitting enemy missile #" + enemy_missile.getID() + " in " + destruct_After_Launch + " sec ");
 						Thread.sleep((long) (destruct_After_Launch * 1000));
-						System.out.println(Calendar.getInstance().getTime()+ "\t Iron dome #" + this.id + " successfully intercepted missile #" + id);
-						destroyMissile(id);
+						if (enemy_missile.isAlive) {
+							War.theLogger.log(Level.INFO, "Iron dome #" + this.id + " successfully intercepted missile #" + enemy_missile.getID(), this);
+							War.theLogger.log(Level.INFO, "Iron dome #" + this.id + " successfully intercepted missile #" + enemy_missile.getID(), enemy_missile.getFather());
+							System.out.println(Calendar.getInstance().getTime() + "\t Iron dome #" + this.id + " successfully intercepted missile #" + id);
+							destroyMissile(id);
+						}
 					}
 				}
 			}
@@ -65,7 +58,9 @@ public class Iron_Dome extends Thread {
 			}
 			synchronized (this) {
 				try {
-					//cause the missile to wait till there are other action, but doesn't need notify because he will end his life anyway
+					// cause the missile to wait till there are other action,
+					// but doesn't need notify because he will end his life
+					// anyway
 					wait();
 				} catch (InterruptedException e) {
 					e.printStackTrace();
@@ -75,9 +70,9 @@ public class Iron_Dome extends Thread {
 	}
 
 	private void setISalive(boolean bool) {
-		this.isAlive = bool;	
+		this.isAlive = bool;
 	}
-	
+
 	public String getDestructAfterLaunch() {
 		return destructAfterLaunch;
 	}
